unified signtare for transform method in FTS apps
